+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.1.2 [unreleased]
 
+- Improved duplicate index logic to detect more duplicates
 - Fixed error with large number of sequences
 
 ## 2.1.1